server/pyproject.toml
[tool.poetry]
name = "alude-todo-server"
version = "0.1.0"
description = "Alude's welcome project server"
authors = ["Gabriel de Paula"]
license = "Unlicense"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10"
quart = ">=0.20.0,<0.21.0"
python-dotenv = "^1.2.1"
hypercorn = "^0.18.0"
sqlalchemy = "^2.0.46"
pydantic = {extras = ["email"], version = "^2.12.5"}
asyncpg = "^0.31.0"
bcrypt = "^5.0.0"
quart-schema = "^0.23.0"
quart-cors = "^0.8.0"
pyjwt = "^2.11.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=9.0.2,<10.0.0"
black = "^26.1.0"

[tool.black]
line-length = 80
target-version = ["py311"]
include = '\.pyi?$'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"